feat(consensus): Optimization for processing blocks and compression for network (#107)
We have added two improvements:

We have added a filter for all connection tasks that allows to not process block/header/shard when it is not needed. Previously this was done by reading the dag state which could be updated not immediately due to the queue in the core task. Now the filter exists outside the dag state and allows for better filtering of unnecessary data.
We have added a zstd compression on the network level. It allows compressing block headers in the bundle of blocks that are sent by validators. This significantly decreases the network bandwidth, especially for a low load.